feat : spring api 연결을 위한 코드변경 및 헤더에 토큰 저장 기능 추가

+++ JS/community.js
@@ -1,21 +1,48 @@
+function get_cookie(name) {
+    let value = document.cookie.match('(^|;) ?' + name + '=([^;]*)(;|$)');
+    return value? value[2] : null; }
+
+
 // 페이지 접속 시 초기화
-$(document).ready(function() {
-    communtityPostsGet();
+$(window.document).ready(function() {
+    communityPostsGet();
 })
 
-// 게시물 POST
-function communtityMakePosts() {
-    let title = $('title_box').val()
-    let content = $('content_box').val()
-    let created_at = new Date().toISOString()
-
+//게시물 GET
+function communityPostsGet() {
+    let token = get_cookie("X-AUTH-TOKEN");
+    console.log(token);
     $.ajax({
-        type: "POST",
-        url: "/user/community/post",
-        data: {title: title, content: content, created_at: created_at},
-        success: function (response) {
-            alert(response['msg']);
-            window.location.reload();
+        type: "GET",
+        url: "user/community/posts",
+        data: {},
+        contentType: "application/json;",
+        beforeSend: function (xhr) {
+            xhr.setRequestHeader("Content-type","application/json");
+            xhr.setRequestHeader("X-AUTH-TOKEN", token);
+        },
+        success: function (posts) {
+            console.log(posts)
+            let post = posts
+            for (let i = 0; i < post.length; i++) {
+                const post_id = post[i]['postId']
+                const title = post[i]['title']
+                const nickname = post[i]['nickname']
+                const create_at = post[i]['createdAt']
+                const time_befor = time2str(create_at)
+                const temp_html = `<div>
+                                    <a class="posting_box" href="${post_id}">
+                                        <p style="font-size: 20px; float: left;">${title}</p>
+                                        <div class="time_box">
+                                            <p class="posting_time">${time_befor}</p>
+                                        </div>
+                                        <div class="time_box" style="text-align: center;">
+                                            <p>${nickname}</p>
+                                        </div>        
+                                    </a>
+                                </div>`
+                $('#community_post').append(temp_html)
+            }
         }
     })
 }
@@ -39,51 +66,41 @@ function time2str(date) {
     return `${date.getFullYear()}년 ${date.getMonth() + 1}월 ${date.getDate()}일`
 }
 
-//게시물 GET
-function communtityPostsGet() {
-    //$('#communtity_post').empty()
+// 게시물 POST
+function communityMakePost() {
+    let title = $('title_box').val()
+    let content = $('content_box').val()
+    let created_at = new Date().toISOString()
+
     $.ajax({
-        type: "GET",
-        url: "/user/community/posts",
-        data: {},
-        success: function (posts) {
-            let post = posts['postId']
-            for (let i = 0; i < post.length; i++) {
-                const post_id = post[i]['postId']
-                const title = post[i]['title']
-                const nickname = post[i]['nickname']
-                const create_at = card[i]['createdAt']
-                const time_befor = time2str(create_at)
-                const temp_html = `<div>
-                                        <a class="posting_box" href="${post_id}">
-                                            <p style="font-size: 20px; float: left;">${title}</p>
-                                            <div class="time_box">
-                                                <p class="posting_time">${time_befor}</p>
-                                            </div>
-                                            <div class="time_box" style="text-align: center;">
-                                                <p>${nickname}</p>
-                                            </div>        
-                                        </a>
-                                    </div>`
-                                $('#communtity_post').append(temp_html)
-            }
+        type: "POST",
+        url: "/user/community/post",
+        data: {title: title, content: content, created_at: created_at},
+        contentType: "application/json;",
+        success: function (response) {
+            console.log(response)
+            alert(response['저장되었습니다.']);
+            window.location.reload("/community.js");
         }
     })
 }
 
 // 게시물 DELETE
-$(document).ready(function() {
-    $('#post_delete').click(function() {
+function community_post_delete(){
+    if(confirm('삭제하겠습니까?')){
         $.ajax({
-        type: "DELETE",
-        url: "/user/community/post/{postId}",
-        data: {},
-        success: function (result) {
-            console.log(result);
-        }
-    })
-    })
-})
+            type: "DELETE",
+            url: "/user/community/post/{postId}",
+            data: {},
+            success: function (result) {
+                console.log(result);
+            }
+        })
+        alert("삭제되었습니다.")
+    } else {
+    }
+}
+
 
 //게시물 수정
 function communityPutGet(post_id) {
@@ -106,8 +123,8 @@ function communityPutGet(post_id) {
                                     <hr class="hr_top">
                                     <textarea id="content_box" placeholder="내용을 입력하세요">${content}</textarea>
                                 </div>`
-                            $('#communitys').append(temp_html)
-            
+            $('#communitys').append(temp_html)
+
         }
     })
 }
